- getMoreInfoAboutFilm
