+++ src/api.ts
@@ -1,5 +1,4 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-import axios, { AxiosInstance } from 'axios';
+import axios from 'axios';
 
 const BASE_URL = 'https://13.design.pages.academy/wtw';
 const API_TIMEOUT = 5000;
@@ -13,14 +12,3 @@ export function createAPI() {
   return api;
 }
 
-export const getFilms = createAsyncThunk('films/getFilms', async (_, { extra: api }) => {
-  const apiInstance = api as AxiosInstance;
-  const response = await apiInstance.get('/films');
-  return response.data; // lint возмущается, что any, unknown делать нельзя, что за данные - не знаю. Эх! :с
-});
-
-export const getFilm = createAsyncThunk('films/getFilm', async (filmId: string, { extra: api }) => {
-  const apiInstance = api as AxiosInstance;
-  const response = await apiInstance.get(`/films/${filmId}`);
-  return response.data;
-});
