+++ src/index.tsx
@@ -4,7 +4,8 @@ import {App} from './components/app';
 import { Detail } from './mocks/details';
 import store from './store';
 import { Provider } from 'react-redux';
-import { getMoreInfoAboutFilm } from './components/functions/get-more-info-about-film';
+// import { getMoreInfoAboutFilm } from './components/functions/get-more-info-about-film';
+import { getFilm } from './store/api-action';
 
 
 const root = ReactDOM.createRoot(
@@ -13,8 +14,12 @@ const root = ReactDOM.createRoot(
 
 //Пока нет авторизации, не могу получить промо-фильм. Беру первый (про Гриндевальда)
 
+
 const state = store.getState();
-const [detailPromoFilm, ] = await getMoreInfoAboutFilm(state.films[0]);
+
+const loadingdetail = await store.dispatch(getFilm(state.films[0].id));
+
+const detailPromoFilm = loadingdetail.payload as Detail;
 const promoFilm = state.films[0];
 
 root.render(
@@ -22,7 +27,7 @@ root.render(
     <Provider store={store}>
       <App paramsMain={{
         film: promoFilm ,
-        detail: detailPromoFilm as Detail
+        detail: detailPromoFilm
       }}
       />
     </Provider>
