--- src/mocks/overview.ts
@@ -1,9 +1,0 @@
-export type Overview = {
-    filmId: string;
-    description: string;
-    rating: number;
-    ratingDescription: string;
-    votes: number;
-    director: string;
-    actors: string[];
-}
